refactor: Move wrongly placed api_failure tests into api_fail.rs
