+++ cloudflare/src/endpoints/zones/zone.rs
@@ -197,7 +197,7 @@ pub struct Zone {
     /// Available permissions on the zone for the current user requesting the item
     pub permissions: Vec<String>,
     /// A zone plan
-    // TODO: Right, but undocumented in the official API docs nor in the official TypeScript library. What should we do?
+    // TODO: Correct, but undocumented in the official API docs nor in the official TypeScript library. What should we do?
     pub plan: Option<Plan>,
     /// A zone plan
     pub plan_pending: Option<Plan>,
