+++ cloudflare/src/framework/response/api_fail.rs
@@ -1,9 +1,9 @@
+use crate::framework::response::ResponseInfo;
 use serde::{Deserialize, Serialize};
 use serde_json::value::Value as JValue;
 use std::collections::HashMap;
 use std::error::Error;
 use std::fmt::{self, Debug, Write as _};
-use crate::framework::response::ResponseInfo;
 
 /// Note that APIErrors's `eq` implementation only compares `code` and `message`.
 /// It does NOT compare the `other` values.
@@ -64,5 +64,47 @@ impl fmt::Display for ApiFailure {
 impl From<reqwest::Error> for ApiFailure {
     fn from(error: reqwest::Error) -> Self {
         ApiFailure::Invalid(error)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::framework::response::ResponseInfo;
+    use std::collections::HashMap;
+
+    #[test]
+    fn api_failure_eq() {
+        let err1 = ApiFailure::Error(
+            reqwest::StatusCode::NOT_FOUND,
+            ApiErrors {
+                errors: vec![ResponseInfo {
+                    code: 1000,
+                    message: "some failed".to_owned(),
+                    other: HashMap::new(),
+                }],
+                other: HashMap::new(),
+            },
+        );
+        assert_eq!(err1, err1);
+
+        let err2 = ApiFailure::Error(
+            reqwest::StatusCode::NOT_FOUND,
+            ApiErrors {
+                errors: vec![ResponseInfo {
+                    code: 1000,
+                    message: "some different thing failed".to_owned(),
+                    other: HashMap::new(),
+                }],
+                other: HashMap::new(),
+            },
+        );
+        assert_ne!(err2, err1);
+
+        let not_real_website = "notavalid:url.evena little";
+        let fail = ApiFailure::Invalid(reqwest::blocking::get(not_real_website).unwrap_err());
+        assert_eq!(fail, fail);
+        assert_ne!(fail, err1);
+        assert_ne!(fail, err2);
     }
 }
